+++ src/rules/media-feature-no-missing-punctuation/__tests__/index.js
@@ -61,6 +61,9 @@ testRule(rule, {
   }, {
     code: "@media screen and (max-width:($var - 42px)) {}",
     description: "ignore non-standard complex SCSS values",
+  }, {
+    code: "@media (max-width: $content-size + $side * 2) {}",
+    description: "ignore non-standard complex SCSS values",
   } ],
 
   reject: [ {